some refactors + adjustments
