+++ frontend/language/src/nb.json
@@ -1510,7 +1510,7 @@
   "ux_editor.component_properties.subform.create_new_data_model_label": "Navn på ny datamodell",
   "ux_editor.component_properties.subform.created_layout_set_name": "Navn på underskjema",
   "ux_editor.component_properties.subform.data_model_binding_label": "Velg datamodellknytning",
-  "ux_editor.component_properties.subform.data_model_empty_messsage": "Ingen tilgjengelige datamodeller",
+  "ux_editor.component_properties.subform.data_model_empty_message": "Ingen tilgjengelige datamodeller",
   "ux_editor.component_properties.subform.go_to_layout_set": "Gå til utforming av underskjemaet",
   "ux_editor.component_properties.subform.layout_set_is_missing_content_heading": "Underskjemaet ditt mangler innhold.",
   "ux_editor.component_properties.subform.layout_set_is_missing_content_paragraph": "Denne tabellen bruker underskjemaet for å hente feltene og tekstene som skal vises i tabellen. Velg Utform underskjemaet for å legge inn innhold.",
